- updated enqueue classes; - updated admin-navmenu class (removed support WordPress < 5.4);
